Refactor and add more tests for gossip fetcher
